fix: ensure create uses entity from options
When an entity is provided in the options to create
it will use that entity as the new object.
Sending it to the server in the post,
but also updating it with the details returned from the post.

This ensures that users can have a pre-created entity
that is then saved to the server.
All values bound to the existing entity will then
also receive details from the post.